mcp-server/requirements.txt
mcp-python-sdk>=0.5.0
moviepy==1.0.3
pydub==0.25.1
openai>=1.12.0
loguru>=0.7.0
